geospatial_dashboard/requirements.txt
fastapi
uvicorn
streamlit
streamlit-folium
folium
requests
numpy
pandas
python-dotenv
